python/topic-suggestion/pyproject.toml
[tool.poetry]
name = "topic-suggestion"
version = "0.1.0"
description = ""
authors = ["THECODEDEMOGORGON <matnik2001@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
edgedb = "^1.9.0"
fastbm25 = "^0.0.2"
requests = "^2.31.0"
bs4 = "^0.0.2"
python-dotenv = "^1.0.1"
uvicorn = "^0.29.0"
fastapi = "^0.111.0"
six = "^1.16.0"
numpy = "^1.26.4"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
coverage = "^7.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "topic-suggestion.main:start"